Update: refactoring brand
