+++ src/main/kotlin/com/musinsa/musinsacodingassignment/brand/domain/Brand.kt
@@ -1,7 +1,5 @@
 package com.musinsa.musinsacodingassignment.brand.domain
 
-import com.musinsa.musinsacodingassignment.brand.entity.BrandEntity
-
 data class Brand(
     var id: Long?,
     var name: String
@@ -11,13 +9,3 @@ data class Brand(
         return this
     }
 }
-
-fun Brand.toEntity() = BrandEntity(
-    id = this.id ?: 0,
-    name = this.name
-)
-
-fun BrandEntity.toDomain() = Brand(
-    id = this.id,
-    name = this.name,
-)
