+++ src/main/kotlin/com/musinsa/musinsacodingassignment/brand/presentation/dto/request/UpdateBrandRequest.kt
@@ -6,7 +6,7 @@ data class UpdateBrandRequest(
     val name: String
 )
 
-fun UpdateBrandRequest.toBrand(id: Long) = Brand(
+fun UpdateBrandRequest.toDomain(id: Long) = Brand(
     id = id,
     name = this.name,
 )
