+++ src/main/kotlin/com/musinsa/musinsacodingassignment/product/application/PriceService.kt
@@ -1,7 +1,7 @@
 package com.musinsa.musinsacodingassignment.product.application
 
-import com.musinsa.musinsacodingassignment.brand.domain.toDomain
 import com.musinsa.musinsacodingassignment.brand.entity.BrandEntity
+import com.musinsa.musinsacodingassignment.brand.entity.toDomain
 import com.musinsa.musinsacodingassignment.category.entity.CategoryEntity
 import com.musinsa.musinsacodingassignment.brand.repository.BrandRepository
 import com.musinsa.musinsacodingassignment.category.entity.toDomain
@@ -81,7 +81,11 @@ class PriceService(
         var resultProductEntities = mutableListOf<ProductEntity>()
 
         brandEntities.forEach { brandEntity ->
-            val (totalCategoryPrice, totalCategoryProductEntities) = calculateTotalCategoryPrice(brandEntity, categoryEntities, allProducts)
+            val (totalCategoryPrice, totalCategoryProductEntities) = calculateTotalCategoryPrice(
+                brandEntity,
+                categoryEntities,
+                allProducts
+            )
             if (totalCategoryPrice != 0 && totalCategoryPrice < resultPrice) {
                 resultBrandEntity = brandEntity
                 resultPrice = totalCategoryPrice
